requirements.txt
# Core Dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
pyarrow>=12.0.0
Pillow>=9.5.0
PyYAML>=6.0
tqdm>=4.65.0

# Optional: Advanced models
# timm>=0.9.0  # For various vision models
# open-clip-torch>=2.20.0  # For OpenCLIP models

# Development dependencies (optional)
# pytest>=7.3.0
# black>=23.3.0
# flake8>=6.0.0
# mypy>=1.3.0

